elyxproject/requirements.txt
crewai>=0.28.8
crewai[tools]>=0.1.6
python-dotenv>=1.0.0
PyYAML>=6.0
langchain-google-genai>=1.0.3
langchain-community>=0.2.19 